requirements.txt
streamlit==1.2.0
pandas==1.3.4
numpy==1.20.3
pillow==8.4.0
scikit-learn==0.24.2